requirements.txt
aiofiles==23.2.1
alembic==1.12.0
annotated-types==0.7.0
anthropic==0.52.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.2
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.1
email-validator==2.0.0
exceptiongroup==1.3.0
fastapi==0.104.1
greenlet==3.1.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
jiter==0.9.1
mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.24.4
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.0.3
pygments==2.19.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
regex==2024.11.6
requests==2.32.3
rich==13.6.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.40
sse-starlette==1.6.5
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.7.0
tqdm==4.66.1
typing-extensions==4.13.2
urllib3==2.2.3
uvicorn==0.23.2
zipp==3.20.2
psycopg2-binary==2.9.9
sentence-transformers==2.2.2
torch>=1.9.0
transformers>=4.21.0
PyMuPDF==1.23.26
pdfminer.six==20231228
